More steps towards cleaning up membership registration
* deal with deprecations warning from the latest material ui upgrade
* start transitioning the state managed dialogs over to route managed ones, so far membership and game edit/creation.
* Fix up room quantity calculations and display, for registering memberships and in the hotel room lists.
* Add a debug button to the wizard - admin only.
* Change text field format from the default "almost invisible" over to the "shaded background" variant.  This actually follows the Material Design convention better than the mui defaults do (and they are going to change it in the next revision).

+++ support/db/migrations/20210720203000_fix_policies.js
@@ -42,9 +42,12 @@ exports.up = async function (knex) {
   const user = process.env.DATABASE_USER
   const password = process.env.DATABASE_USER_PASSWORD ?? ''
 
-  await knex.raw(`drop role if exists ${user}`)
-  await knex.raw(`CREATE ROLE ${user} WITH LOGIN PASSWORD '${password}' NOINHERIT;`)
-  // await knex.raw(``)
+  const res = knex.raw(`SELECT 1 FROM pg_roles WHERE rolname=${user}`)
+  if (res !== '1') {
+    // note that users are per database and not per schema
+    await knex.raw(`drop role if exists ${user}`)
+    await knex.raw(`CREATE ROLE ${user} WITH LOGIN PASSWORD '${password}' NOINHERIT;`)
+  }
   await knex.raw(`grant usage on schema public to ${user};`)
   await knex.raw(`grant select, insert, update, delete on all tables in schema public to ${user};`)
   await knex.raw(`grant select, update, usage on all sequences in schema public to ${user};`)
